Papp 422 | Fix v8 upgrade issues (#173)
* Fix orientation issue in 8.1.0

* Upgrade to v8.6.0

* Adding few logs
