+++ pom.xml
@@ -43,7 +43,7 @@
     <dependency>
         <groupId>io.appium</groupId>
         <artifactId>java-client</artifactId>
-        <version>8.1.0</version>
+        <version>8.6.0</version>
     </dependency>
     <dependency>
       <groupId>org.json</groupId>
