+++ src/main/java/io/percy/appium/metadata/IosMetadata.java
@@ -9,11 +9,13 @@ import io.percy.appium.lib.Cache;
 public class IosMetadata extends Metadata {
     private IOSDriver driver;
     private String sessionId;
+    private String orientation;
 
     public IosMetadata(AppiumDriver driver, String deviceName, Integer statusBar, Integer navBar, String orientation,
             String platformVersion) {
         super(driver, deviceName, statusBar, navBar, orientation, platformVersion);
         this.driver = (IOSDriver) driver;
+        this.orientation = orientation;
         this.sessionId = driver.getSessionId().toString();
     }
 
@@ -81,5 +83,29 @@ public class IosMetadata extends Metadata {
     public Integer scaleFactor() {
         return Integer.valueOf(getSession().get("pixelRatio").toString());
     }
+
+
+    public String orientation() {
+      if (orientation != null) {
+          if (orientation.toLowerCase().equals("portrait") || orientation.toLowerCase().equals("landscape")) {
+              return orientation.toLowerCase();
+          } else if (orientation.toLowerCase().equals("auto")) {
+              try {
+                  return driver.getOrientation().toString().toLowerCase();
+              } catch (java.lang.NoSuchMethodError e) {
+                  return "portrait";
+              }
+          } else {
+              return "portrait";
+          }
+      } else {
+          Object orientationCapability = driver.getCapabilities().getCapability("orientation");
+          if (orientationCapability != null) {
+              return orientationCapability.toString().toLowerCase();
+          } else {
+              return "portrait";
+          }
+      }
+  }
 
 }
