+++ src/main/java/io/percy/appium/metadata/Metadata.java
@@ -47,29 +47,6 @@ public abstract class Metadata {
         return osVersion.toString();
     }
 
-    public String orientation() {
-        if (orientation != null) {
-            if (orientation.toLowerCase().equals("portrait") || orientation.toLowerCase().equals("landscape")) {
-                return orientation.toLowerCase();
-            } else if (orientation.toLowerCase().equals("auto")) {
-                try {
-                    return driver.getOrientation().toString().toLowerCase();
-                } catch (java.lang.NoSuchMethodError e) {
-                    return "portrait";
-                }
-            } else {
-                return "portrait";
-            }
-        } else {
-            Object orientationCapability = driver.getCapabilities().getCapability("orientation");
-            if (orientationCapability != null) {
-                return orientationCapability.toString().toLowerCase();
-            } else {
-                return "portrait";
-            }
-        }
-    }
-
     public String getDeviceName() {
         return deviceName;
     }
@@ -92,6 +69,8 @@ public abstract class Metadata {
     }
 
     public abstract Integer deviceScreenWidth();
+
+    public abstract String orientation();
 
     public abstract String deviceName();
